Factor out login-form.tsx.
